Fix build on MacOS
`make install` executes `ocamlfind install -patch-version $(VERSION)`. The VERSION
script used `echo -n` to output the version number without a trailing newline,
but echo does not support -n on MacOS (Sierra). Because the -n option is not
necessary (bash will convert newlines to whitespace when tokenizing the output
of a command in the context of command substitution), I am removing the -n
option from VERSION.